Fix gitlab-shell issue
This commit fixes an issue where gitlab-shell wouldn't work after
starting gitlab. The issue was gitlab was running on port 443, while the
shell was trying to access it locally over port 80. This commit updates
the config file for gitlab-shell to use 127.0.0.1:80 to talk with
gitlab, while nginx is listening locally on port 80 for this connection.