Fix startup scripts
This commit fixes the startup script for gitlab and adds a call to
remove the socket gitlab uses before trying to start up. It also fixes
the gitlab-start and sidekiq execute blocks to not be user specific
since we are handling this in the command. It also fixes a call to fqdn
and ensures we grab this value from the node. And it also removes the
default nginx site configuration from sites-enabled since we want gitlab
to be the only site that nginx serves by default.