Add attribute files
This commit adds separate attribute files for all of the different
attributes, based on similarly named recipes. This is because it is chef
convention to do so. It also starts implementing the backup recipe. And
it also updates the dependencies in the Gemfile and Berksfile. And
finally it updates the vagrant file to be set up correctly for deploying
this cookbook on ubuntu.