Update strainerfile and travis config
This commit adds back in the commented out lines for foodcritic and the
cookbook test. It also updates the travis yaml to run strainer as this
is now the preferred method to run tests on this cookbook. It also adds
additional specs for the cookbook. And it removes the Thorfile as it is
not needed or used.