Add git dependency for cookbook
This commit adds the git dependency for the cookbook. Since we aren't
including the nginx cookbook, we don't have access to the nginx resource
to restart it. Also it fixes a typo where we include the postgresql
recipe from the database cookbook. And it also adds the git source recipe
so we compile git from source instead of installing a package.