Set git prefix attribute
This commit adds a normal attribute for git prefix path. This allows us
to change where git is installed. This commit also starts the spec for
gitlab_shell. And finally this commit starts an implementation of a
custom rspec matcher for a git resource. This commit also updates the
readme.