Add cookbooks to Berksfile
This commit adds the ulimit cookbook to the Berksfile. This is because
there is an error when using a newer version of ulimit cookbook with
chef-solo. It also adds the sudo cookbook to the Berksfile. This is
because when running the vagrant vm, after the first run the vagrant
user loses sudo privileges. This is used to ensure vagrant has sudo
privileges.