Add attributes to default attributes
This commit updates the git version to 1.8.4. This is because the older
version of git had a bug when using the git resource to checkout files
as a particular user. By upgrading, we fix the issue. It also adds
build-essential compiletime true. This ensures that build-essential is
installed immediately when included in the run list. It also includes
nginx default_site_enabled false to prevent nginx from outputting the
default site configuration. This is because we only want the gitlab conf
to be rendered.