Add gitlab service notifiers
This commit adds a gitlab service resource. It also adds notifiers to
the gitlab template files so it notifies the service to restart. It also
includes a notifier for nginx when the website configuration file
changes. It also removes an unneeded execute resource block that was
used to set git config values. However the resource block wasn't needed
because there was already two execute resource blocks in place to do the
same. And it also changes the restart function in the gitlab init
script. Instead of exiting with a status of 1, it calls start on the
gitlab service.